+++ src/app/api/posts/[slug]/route.ts
@@ -1,5 +1,6 @@
 import prisma from '@/db/connect'
 import { NextResponse } from 'next/server'
+import { getAuthSession } from '../../auth/[...nextauth]/options'
 
 interface ExtendedPostUpdateInput {
   views: {
@@ -36,7 +37,11 @@ export const DELETE = async (
   { params }: { params: { slug: string } },
 ) => {
   const { slug } = params
+  const session = await getAuthSession()
 
+  if (!session) {
+    return new NextResponse(JSON.stringify({ message: 'Not Authenticated!' }))
+  }
   try {
     // Delete the post by slug
     const deleteResult = await prisma.post.delete({
